Copybara import of the project:
--
226520691 by Googler <noreply@google.com>:

Hey so this is unrelated to a current bug, but we want to be able to get and set the type of our shadow bluetooth device.

Because this is currently not implemented we have some test that cannot go through the full flow and gets around this by just calling methods directly, which can be flaky.

PiperOrigin-RevId: 226520691